+++ src/command_modules/azure-cli-acs/azure/cli/command_modules/acs/custom.py
@@ -53,10 +53,10 @@ from azure.mgmt.containerservice.models import ContainerServiceServicePrincipalP
 from azure.mgmt.containerservice.models import ContainerServiceSshConfiguration
 from azure.mgmt.containerservice.models import ContainerServiceSshPublicKey
 from azure.mgmt.containerservice.models import ContainerServiceStorageProfileTypes
-from azure.mgmt.containerservice.models import ManagedCluster
-from azure.mgmt.containerservice.models import ManagedClusterAADProfile
-from azure.mgmt.containerservice.models import ManagedClusterAddonProfile
-from azure.mgmt.containerservice.models import ManagedClusterAgentPoolProfile
+from azure.mgmt.containerservice.v2018_03_31.models import ManagedCluster
+from azure.mgmt.containerservice.v2018_03_31.models import ManagedClusterAADProfile
+from azure.mgmt.containerservice.v2018_03_31.models import ManagedClusterAddonProfile
+from azure.mgmt.containerservice.v2018_03_31.models import ManagedClusterAgentPoolProfile
 from ._client_factory import cf_container_services
 from ._client_factory import cf_resource_groups
 from ._client_factory import get_auth_management_client
@@ -1611,6 +1611,22 @@ def aks_list(cmd, client, resource_group_name=None):
 def aks_show(cmd, client, resource_group_name, name):
     mc = client.get(resource_group_name, name)
     return _remove_nulls([mc])[0]
+
+
+def aks_update_credentials(cmd, client, resource_group_name, name, 
+               reset_service_principal=False,
+               service_principal=None,
+               client_secret=None,
+               no_wait=False):
+    if reset_service_principal != 1:
+        raise CLIError('Please specify "--reset-service-principal".')
+    
+    if service_principal is None or client_secret is None:
+        raise CLIError('Please specify --service-principal and --client-secret when --reset-service-principal flag is on.')
+    return sdk_no_wait(no_wait, 
+                        client.reset_service_principal_profile,
+                        resource_group_name,
+                        name, service_principal, client_secret)
 
 
 def aks_scale(cmd, client, resource_group_name, name, node_count, nodepool_name="", no_wait=False):